Adapting changes to updated project structure merged from master.
