+++ src/views/MainScreen.tsx
@@ -3,10 +3,14 @@ import {
 } from 'react-native';
 import React, { useState } from 'react';
 import {
-  ExerciseRecord, usePostExerciseRecord, useGetCounter, useGetExerciseRecords
+  ExerciseRecord, usePostExerciseRecord, useGetExerciseRecords
 } from '../api-client/api-client';
 import PocButton from '../components/PocButton';
 import PocPrompt from '../components/PocPrompt';
+
+import getEnvVars from '../utils/environment';
+
+const { message } = getEnvVars();
 
 const styles = StyleSheet.create({
   container: {
@@ -23,19 +27,13 @@ const styles = StyleSheet.create({
   }
 });
 
-interface MainProps{
-    message: string;
-}
-
-export default function Main(props: MainProps) {
-  const { message } = props;
+export default function MainScreen() {
   const userId = 1;
   const [exerciseLabel, exerciseLabelUpdate] = useState('My exercise');
   const [exerciseCalories, exerciseCaloriesUpdate] = useState('100');
 
   const tomorrow = new Date();
   tomorrow.setDate(tomorrow.getDate() + 1);
-  const { data: counter, refetch: refetchCounter } = useGetCounter({});
   const { data: exercises, refetch: refetchExercises } = useGetExerciseRecords({ userId, queryParams: { endDateTime: tomorrow.toISOString(), startDateTime: '0000-00-00 00:00:00' } });
   const { mutate: postExercise } = usePostExerciseRecord({ userId });
 
@@ -58,12 +56,6 @@ export default function Main(props: MainProps) {
   return (
     <View style={styles.container}>
       <Text>{message}</Text>
-      <Text>
-        Counter:
-        {' '}
-        {counter && counter.Value }
-      </Text>
-      <PocButton title="Click Me" onPress={() => refetchCounter()} />
       <PocPrompt />
 
       <ScrollView>
